Split JS into separate files
